fix: adapt codebase to qmetaobject 0.2.10 API changes
- qrc! macro: new syntax with comma after function name and 'base_dir as prefix' format
- #[qt_method] attribute → qt_method!() function-like macro declarations in struct
- register_type → qml_register_type, register_singleton_type → qml_register_singleton_type
- qml_register_* now uses cstr!() and no &mut engine parameter
- Added cstr = 0.2 dependency
- qmetaobject::Singleton trait → inherent get() method + QSingletonInit trait impl
- QAbstractListModel::row_count() now returns i32 instead of usize
- log::warn! → ::log::warn! to disambiguate from qmetaobject glob import
- Added missing QPointer/QString imports in file_transfer.rs
- ImageInfo::new() → ImageInfo::default() (constructor removed in ruma 0.16)
- Removed unused Arc/Mutex imports in profile.rs

+++ src/room_model.rs
@@ -113,8 +113,8 @@ impl RoomModel {
 
 // QAbstractListModel glue.
 impl qmetaobject::QAbstractListModel for RoomModel {
-    fn row_count(&self) -> usize {
-        self.entries.borrow().len()
+    fn row_count(&self) -> i32 {
+        self.entries.borrow().len() as i32
     }
     fn data(&self, index: qmetaobject::QModelIndex, role: i32) -> qmetaobject::QVariant {
         let i = index.row() as usize;
