+++ src/spaces.rs
@@ -182,8 +182,8 @@ impl SpaceModel {
 }
 
 impl qmetaobject::QAbstractListModel for SpaceModel {
-    fn row_count(&self) -> usize {
-        self.entries.borrow().len()
+    fn row_count(&self) -> i32 {
+        self.entries.borrow().len() as i32
     }
     fn data(&self, index: qmetaobject::QModelIndex, role: i32) -> qmetaobject::QVariant {
         let i = index.row() as usize;
